requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pymupdf>=1.18.0
numpy>=1.21.0
python-multipart